Day 2.5 of working on this project. (Change log bellow)
December/23/2021:

-New Dependencies:


This file is now *optionally* dependent on another of my files. along with the new dependency comes with the ability to display the following:

>Integral Area Illustration(a graph with an integral equal to the original but in an easier format for understanding { A=(lim1-lim0)*A/(lim1-lim0) } )
>Absolute integral value(the printed integral value)
>Function (the section of the function inputted in )

The file will most likely be available in this folder shortly

-Significant Changes:


I've experimented with adding a "base_y_plane" to determine the min y value before it is registered as part of the area but have since determined it to be unnecesary and mildly confusing to use for new users. If you looked forward to that (as if you'd look forward to this at all assuming you even read this) then you can always either implement it yourself or change the derivative.


-Minor Changes:

>Changes in some variable names
>When initializing the "function" class the parameter "func" has been changed to "derivative" to better fit its purpose
>Other minor changes